requirements.txt
fastapi
uvicorn
python-dotenv
fal-client
streamlit
python-multipart
pydantic 